backend/requirements.txt
Flask==2.3.2
redis==4.5.5
psycopg2-binary==2.9.6
